Moved Atrium asserts class to appropriate folder
